refactor: simplify deck start session setup
