+++ src/features/study-session-start/index.ts
@@ -1,2 +1,2 @@
-export { useStartStudySession } from "./model/useStartStudySession";
+export { startStudy } from "./model/startStudy";
 export { StudySessionStartView } from "./ui/StudySessionStartView";
